backend/requirements.txt
python-dotenv
requests
sentence-transformers
transformers
weaviate-client
langchain
langchain-community
langchain-huggingface
langchain-google-genai
langchain-weaviate
